重构审批流程添加 修改部分问题

+++ approve/src/main/java/com/unique/approve/service/impl/ExamineServiceImpl.java
@@ -1,20 +1,25 @@
 package com.unique.approve.service.impl;
 
+import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.util.ObjectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.unique.approve.entity.bo.ExamineSaveBO;
 import com.unique.approve.entity.bo.ExamineSearchBO;
 import com.unique.approve.entity.po.*;
 import com.unique.approve.entity.vo.ExamineVO;
+import com.unique.approve.enums.ExamineStatusEnum;
 import com.unique.approve.mapper.ExamineMapper;
 import com.unique.approve.service.*;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.unique.approve.utils.ApproveUtil;
 import com.unique.core.common.BasePage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * <p>
@@ -31,6 +36,13 @@ public class ExamineServiceImpl extends ServiceImpl<ExamineMapper, Examine> impl
     private IExamineSettingService examineSettingService;
     @Autowired
     private IExamineSettingUserService examineSettingUserService;
+    @Autowired
+    private IExamineNodeService examineNodeService;
+    @Autowired
+    private IExamineNodeUserService examineNodeUserService;
+
+    @Autowired
+    private IExamineRecordService examineRecordService;
 
 
     @Override
@@ -38,20 +50,43 @@ public class ExamineServiceImpl extends ServiceImpl<ExamineMapper, Examine> impl
     public void addOrUpdate(ExamineSaveBO saveBO) {
         //保存
         Examine examine = saveBO.getExamine();
-        Boolean isAdd = ObjectUtil.isEmpty(examine.getId()) ? Boolean.TRUE : Boolean.FALSE;
-        if (isAdd) {
-            if (save(examine)) {
-                saveBO.getExamineSetting().setExamineId(examine.getId());
-                examineSettingService.save(saveBO.getExamineSetting());
-                //用户
-                saveBO.getExamineSettingUser().forEach(r->r.setExamineId(examine.getId()));
-                examineSettingUserService.saveBatch(saveBO.getExamineSettingUser());
-            }
-        }else{
-            updateById(examine);
-            examineSettingService.updateById(saveBO.getExamineSetting());
-            examineSettingUserService.saveOrUpdateBatch(saveBO.getExamineSettingUser());
+        if (ObjectUtil.isNotEmpty(examine.getId())) {
+            lambdaUpdate().set(Examine::getStatus, ExamineStatusEnum.OFF.getType())
+                    .eq(Examine::getId,examine.getId()).update();
         }
+        //添加
+        examine.setId(null);
+        if (save(examine)) {
+            Long examineId = examine.getId();
+            saveBO.getExamineSetting().setExamineId(examineId);
+            //用户
+            saveBO.getExamineSettingUser().forEach(r->r.setExamineId(examineId));
+            //添加子节点
+            List<ExamineNode> examineNodes = ApproveUtil.recursionBuildHadSubList(saveBO.getExamineNodeList(), 0L);
+            List<ExamineNodeUser> nodeUserList = new ArrayList<>();
+            examineNodes.forEach(r->{
+                r.setExamineId(examineId);
+                if (CollectionUtil.isNotEmpty(r.getNodeUserList())) {
+                    r.getNodeUserList().forEach(u->{
+                        u.setExamineId(examineId);
+                    });
+                    nodeUserList.addAll(r.getNodeUserList());
+                }
+            });
+            //保存examinesetting
+            examineSettingService.save(saveBO.getExamineSetting());
+            //保存examinesettingUser
+            examineSettingUserService.saveBatch(saveBO.getExamineSettingUser());
+            //添加节点
+            if (CollectionUtil.isNotEmpty(examineNodes)) {
+                examineNodeService.saveBatch(examineNodes);
+            }
+            //添加节点user
+            if (CollectionUtil.isNotEmpty(nodeUserList)) {
+                examineNodeUserService.saveBatch(nodeUserList);
+            }
+        }
+
     }
 
     @Override
@@ -62,8 +97,13 @@ public class ExamineServiceImpl extends ServiceImpl<ExamineMapper, Examine> impl
     @Override
     @Transactional
     public void deteleByIds(List<Long> ids) {
-        removeByIds(ids);
-        examineSettingService.remove(new LambdaQueryWrapper<ExamineSetting>().in(ExamineSetting::getExamineId,ids));
-        examineSettingUserService.remove(new LambdaQueryWrapper<ExamineSettingUser>().in(ExamineSettingUser::getExamineId,ids));
+        List<ExamineRecord> hadRecordList = examineRecordService.lambdaQuery().in(ExamineRecord::getExamineId, ids).list();
+        if (CollectionUtil.isEmpty(hadRecordList)) {
+            removeByIds(ids);
+            examineSettingService.remove(new LambdaQueryWrapper<ExamineSetting>().in(ExamineSetting::getExamineId,ids));
+            examineSettingUserService.remove(new LambdaQueryWrapper<ExamineSettingUser>().in(ExamineSettingUser::getExamineId,ids));
+            examineNodeService.remove(new LambdaQueryWrapper<ExamineNode>().in(ExamineNode::getExamineId,ids));
+            examineNodeUserService.remove(new LambdaQueryWrapper<ExamineNodeUser>().in(ExamineNodeUser::getExamineId,ids));
+        }
     }
 }
